python/requirements.txt
networkx

nx-arangodb

langgraph

langchain-core
langchain-openai
langchain-community

mysql-connector-python

python-arango